start on the mod split
