+++ listen-kit/Cargo.toml
@@ -7,7 +7,7 @@ description = "Blockchain actions for AI agents"
 documentation = "https://docs.listen-rs.com"
 
 [features]
-default = ["http"]
+default = ["http", "solana"]
 http = [
   "actix-web",
   "actix-cors",
@@ -15,42 +15,49 @@ http = [
   "futures-util",
   "tokio-stream",
 ]
+solana = [
+  "solana-account-decoder",
+  "solana-sdk",
+  "spl-token",
+  "solana-client",
+  "solana-transaction-status",
+  "spl-associated-token-account",
+]
+evm = []
 
 [dependencies]
 # Core dependencies
 env_logger = "0.11.6"
 log = "0.4.25"
-solana-account-decoder = "2.1.9"
-solana-sdk = "2.1.9"
-spl-token = "7.0.0"
 reqwest = { version = "0.12.4", features = ["json", "stream", "multipart"] }
 bincode = "1.3.3"
 borsh = "1.5.1"
 base64 = "0.22.1"
 timed = "0.2.1"
-solana-client = "2.1.9"
 serde = "1.0.199"
 serde_json = "1.0.116"
-solana-transaction-status = "2.1.9"
 rand = "0.8.5"
 chrono = "0.4.39"
 dotenv = "0.15.0"
 hex = "0.4.3"
-tokio = { version = "1.37.0", features = ["full"] }
-spl-associated-token-account = "6.0.0"
-serde_with = "3.12.0"
-proc-macro2 = "1.0.93"
-quote = "1.0.38"
-syn = { version = "1.0", features = ["full"] }
 anyhow = "1.0.95"
 rig-core = { version = "0.6.1", git = "https://github.com/piotrostr/rig", branch = "feat/anthropic-streaming-api" }
-once_cell = "1.20.2"
 rig-tool-macro = "0.2.0"
 thiserror = "2.0.11"
+once_cell = "1.20.2"
+tokio = { version = "1.37.0", features = ["full"] }
 tracing-subscriber = "0.3.19"
 tracing = "0.1.41"
 lazy_static = "1.5.0"
 futures = "0.3.31"
+serde_with = "3.12.0"
+
+solana-account-decoder = { version = "2.1.9", optional = true }
+solana-sdk = { version = "2.1.9", optional = true }
+spl-token = { version = "7.0.0", optional = true }
+solana-client = { version = "2.1.9", optional = true }
+solana-transaction-status = { version = "2.1.9", optional = true }
+spl-associated-token-account = { version = "6.0.0", optional = true }
 
 actix-web = { version = "4", optional = true }
 actix-cors = { version = "0.6", optional = true }
