+++ listen-kit/src/bin/server.rs
@@ -1,7 +1,5 @@
-#[cfg(feature = "http")]
-use listen_kit::http::run_server;
+use listen_kit::http::server::run_server;
 
-#[cfg(feature = "http")]
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
     use listen_kit::agent::create_trader_agent;
